# net/art: implement the path compression optimization