Several individual changes. Commits 2 and 3 are the meat of the optimization's implementation, the rest are small fixups and lots of tests.

This implementation currently passes all tests with -count=1000 reliably, so this _appears_ to be correct code. I still need to run benchmarks again, I'll post those in a comment later.