@danderson 2023-07-13
Benchmarks are below. They construct random route tables of various sizes, and measure performance. Short 
 - IPv4 (IPv6) route insertion is ~30% (~80%) faster, uses ~30% (~80%) less memory.
 - IPv4 (IPv6) route deletion is ~20% slower (~30-60% faster), no change in memory.
 - Route lookup is ~25% slower across the board, although still very fast (12-25 million lookups/sec, depending on test conditions). Memory behavior unchanged.

As expected, the major gain is that IPv6 routes now cost approximately as much IPv4 routes (~10-15kiB per route), instead of consuming 6x more memory for the same amount of information.

There's still opportunities to optimize further (removing the one alloc in the route lookup path is trivial, another more complex optimization could halve the memory cost per route across the board). But assuming this code is correct, it no longer has a completely pathological performance envelope.

```
goos: linux
goarch: amd64
pkg: tailscale.com/net/art
cpu: AMD Ryzen 9 5950X 16-Core Processor
                          │  before2.txt  │              after2.txt              │
                          │    sec/op     │    sec/op     vs base                │
TableInsertion/ipv4/10      1091.5n ±  1%   787.2n ±  1%  -27.88% (p=0.000 n=10)
TableInsertion/ipv4/100     2701.5n ±  4%   923.6n ±  0%  -65.81% (p=0.000 n=10)
TableInsertion/ipv4/1000     2.095µ ±  2%   1.678µ ±  2%  -19.93% (p=0.000 n=10)
TableInsertion/ipv4/10000    2.830µ ±  7%   2.083µ ±  5%  -26.40% (p=0.000 n=10)
TableInsertion/ipv6/10       4.266µ ±  6%   1.112µ ±  1%  -73.93% (p=0.000 n=10)
TableInsertion/ipv6/100     11.154µ ±  2%   1.307µ ±  9%  -88.28% (p=0.000 n=10)
TableInsertion/ipv6/1000    15.263µ ±  4%   2.304µ ±  4%  -84.90% (p=0.000 n=10)
TableInsertion/ipv6/10000   15.166µ ±  5%   2.947µ ±  5%  -80.57% (p=0.000 n=10)
TableDelete/ipv4/10          218.8n ±  2%   292.3n ±  6%  +33.59% (p=0.000 n=10)
TableDelete/ipv4/100         367.3n ±  1%   275.2n ±  2%  -25.07% (p=0.000 n=10)
TableDelete/ipv4/1000        349.5n ±  3%   406.7n ±  2%  +16.35% (p=0.000 n=10)
TableDelete/ipv4/10000       602.8n ±  7%   591.7n ±  5%        ~ (p=0.796 n=10)
TableDelete/ipv6/10          666.8n ±  5%   450.9n ±  5%  -32.37% (p=0.000 n=10)
TableDelete/ipv6/100        1008.5n ±  5%   174.0n ±  2%  -82.75% (p=0.000 n=10)
TableDelete/ipv6/1000       1436.5n ±  5%   507.7n ±  3%  -64.66% (p=0.000 n=10)
TableDelete/ipv6/10000      1700.0n ±  7%   730.6n ± 13%  -57.02% (p=0.000 n=10)
TableGet/ipv4/10             33.63n ±  3%   40.85n ±  2%  +21.47% (p=0.000 n=10)
TableGet/ipv4/100            38.01n ±  2%   50.99n ±  2%  +34.15% (p=0.000 n=10)
TableGet/ipv4/1000           45.09n ±  2%   61.57n ±  2%  +36.56% (p=0.000 n=10)
TableGet/ipv4/10000          50.44n ± 12%   68.38n ± 11%  +35.54% (p=0.000 n=10)
TableGet/ipv6/10             41.80n ±  1%   45.82n ±  1%   +9.62% (p=0.000 n=10)
TableGet/ipv6/100            44.28n ±  1%   57.59n ±  1%  +30.04% (p=0.000 n=10)
TableGet/ipv6/1000           48.82n ±  6%   65.79n ±  3%  +34.76% (p=0.000 n=10)
TableGet/ipv6/10000          67.26n ± 12%   83.48n ±  2%  +24.12% (p=0.000 n=10)
geomean                      509.7n         324.0n        -36.43%

                          │  before2.txt  │               after2.txt               │
                          │   avg-B/op    │   avg-B/op    vs base                  │
TableInsertion/ipv4/10       18.55Ki ± 0%   10.62Ki ± 0%  -42.78% (p=0.000 n=10)
TableInsertion/ipv4/100      19.88Ki ± 0%   11.00Ki ± 0%  -44.65% (p=0.000 n=10)
TableInsertion/ipv4/1000     14.04Ki ± 0%   10.90Ki ± 0%  -22.38% (p=0.000 n=10)
TableInsertion/ipv4/10000   12.904Ki ± 0%   8.995Ki ± 0%  -30.29% (p=0.000 n=10)
TableInsertion/ipv6/10       70.24Ki ± 0%   15.93Ki ± 0%  -77.32% (p=0.000 n=10)
TableInsertion/ipv6/100      94.49Ki ± 0%   13.93Ki ± 0%  -85.26% (p=0.000 n=10)
TableInsertion/ipv6/1000     93.19Ki ± 0%   15.04Ki ± 0%  -83.86% (p=0.000 n=10)
TableInsertion/ipv6/10000    90.96Ki ± 0%   13.19Ki ± 0%  -85.50% (p=0.000 n=10)
TableDelete/ipv4/10            3.200 ± 0%     3.200 ± 0%        ~ (p=1.000 n=10)
TableDelete/ipv4/100           4.030 ± 0%     3.930 ± 0%   -2.48% (p=0.000 n=10)
TableDelete/ipv4/1000          3.990 ± 0%     3.990 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv4/10000         4.000 ± 0%     4.000 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv6/10            16.00 ± 0%     16.00 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv6/100           16.00 ± 0%     15.84 ± 0%   -1.00% (p=0.000 n=10)
TableDelete/ipv6/1000          16.00 ± 0%     15.98 ± 0%   -0.12% (p=0.000 n=10)
TableDelete/ipv6/10000         16.00 ± 0%     16.00 ± 0%        ~ (p=1.000 n=10) ¹
geomean                        545.4          311.7       -42.86%
¹ all samples are equal

                          │  before2.txt  │               after2.txt                │
                          │ avg-allocs/op │ avg-allocs/op  vs base                  │
TableInsertion/ipv4/10         2.400 ± 0%      1.700 ± 0%  -29.17% (p=0.000 n=10)
TableInsertion/ipv4/100        2.500 ± 0%      1.810 ± 0%  -27.60% (p=0.000 n=10)
TableInsertion/ipv4/1000       2.060 ± 0%      1.820 ± 0%  -11.65% (p=0.000 n=10)
TableInsertion/ipv4/10000      1.970 ± 0%      1.680 ± 0%  -14.72% (p=0.000 n=10)
TableInsertion/ipv6/10         6.300 ± 0%      2.100 ± 0%  -66.67% (p=0.000 n=10)
TableInsertion/ipv6/100        8.130 ± 0%      2.040 ± 0%  -74.91% (p=0.000 n=10)
TableInsertion/ipv6/1000       8.030 ± 0%      2.130 ± 0%  -73.47% (p=0.000 n=10)
TableInsertion/ipv6/10000      7.860 ± 0%      1.990 ± 0%  -74.68% (p=0.000 n=10)
TableDelete/ipv4/10            1.000 ± 0%      1.000 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv4/100         1000.0m ± 0%     990.0m ± 0%   -1.00% (p=0.000 n=10)
TableDelete/ipv4/1000          1.000 ± 0%      1.000 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv4/10000         1.000 ± 0%      1.000 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv6/10            1.000 ± 0%      1.000 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv6/100         1000.0m ± 0%     990.0m ± 0%   -1.00% (p=0.000 n=10)
TableDelete/ipv6/1000          1.000 ± 0%      1.000 ± 0%        ~ (p=1.000 n=10) ¹
TableDelete/ipv6/10000         1.000 ± 0%      1.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean                        2.023           1.377       -31.92%
¹ all samples are equal

                          │ before2.txt │               after2.txt               │
                          │  routes/s   │   routes/s     vs base                 │
TableInsertion/ipv4/10      916.3k ± 1%   1270.3k ±  1%   +38.64% (p=0.000 n=10)
TableInsertion/ipv4/100     370.2k ± 4%   1082.8k ±  0%  +192.49% (p=0.000 n=10)
TableInsertion/ipv4/1000    477.4k ± 2%    596.1k ±  2%   +24.86% (p=0.000 n=10)
TableInsertion/ipv4/10000   353.3k ± 7%    480.1k ±  5%   +35.87% (p=0.000 n=10)
TableInsertion/ipv6/10      234.4k ± 6%    899.0k ±  1%  +283.45% (p=0.000 n=10)
TableInsertion/ipv6/100     89.66k ± 2%   765.16k ±  8%  +753.41% (p=0.000 n=10)
TableInsertion/ipv6/1000    65.52k ± 4%   434.02k ±  4%  +562.45% (p=0.000 n=10)
TableInsertion/ipv6/10000   65.94k ± 5%   339.30k ±  5%  +414.59% (p=0.000 n=10)
TableDelete/ipv4/10         4.570M ± 2%    3.421M ±  6%   -25.15% (p=0.000 n=10)
TableDelete/ipv4/100        2.723M ± 1%    3.634M ±  3%   +33.46% (p=0.000 n=10)
TableDelete/ipv4/1000       2.861M ± 3%    2.459M ±  2%   -14.05% (p=0.000 n=10)
TableDelete/ipv4/10000      1.659M ± 7%    1.690M ±  5%         ~ (p=0.796 n=10)
TableDelete/ipv6/10         1.500M ± 6%    2.218M ±  4%   +47.85% (p=0.000 n=10)
TableDelete/ipv6/100        991.6k ± 6%   5746.5k ±  2%  +479.49% (p=0.000 n=10)
TableDelete/ipv6/1000       696.2k ± 5%   1969.7k ±  3%  +182.91% (p=0.000 n=10)
TableDelete/ipv6/10000      588.4k ± 8%   1368.8k ± 12%  +132.61% (p=0.000 n=10)
geomean                     584.8k         1.305M        +123.17%

                    │ before2.txt  │              after2.txt              │
                    │   addrs/s    │   addrs/s     vs base                │
TableGet/ipv4/10      29.74M ±  3%   24.48M ±  2%  -17.68% (p=0.000 n=10)
TableGet/ipv4/100     26.31M ±  2%   19.61M ±  2%  -25.46% (p=0.000 n=10)
TableGet/ipv4/1000    22.18M ±  2%   16.24M ±  2%  -26.77% (p=0.000 n=10)
TableGet/ipv4/10000   19.82M ± 10%   14.63M ± 10%  -26.22% (p=0.000 n=10)
TableGet/ipv6/10      23.92M ±  1%   21.82M ±  1%   -8.78% (p=0.000 n=10)
TableGet/ipv6/100     22.58M ±  1%   17.36M ±  1%  -23.11% (p=0.000 n=10)
TableGet/ipv6/1000    20.48M ±  6%   15.20M ±  3%  -25.80% (p=0.000 n=10)
TableGet/ipv6/10000   14.90M ± 11%   11.98M ±  2%  -19.58% (p=0.000 n=10)
geomean               22.09M         17.26M        -21.88%

                    │ before2.txt │             after2.txt              │
                    │    B/op     │    B/op     vs base                 │
TableGet/ipv4/10       4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/100      4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/1000     4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/10000    4.000 ± 0%   4.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/10       16.00 ± 0%   16.00 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/100      16.00 ± 0%   16.00 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/1000     16.00 ± 0%   16.00 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/10000    16.00 ± 0%   16.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                8.000        8.000       +0.00%
¹ all samples are equal

                    │ before2.txt │             after2.txt              │
                    │  allocs/op  │ allocs/op   vs base                 │
TableGet/ipv4/10       1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/100      1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/1000     1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv4/10000    1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/10       1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/100      1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/1000     1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
TableGet/ipv6/10000    1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                1.000        1.000       +0.00%
¹ all samples are equal
```